requirements.txt
gradio
ultralytics
torch
torchvision
pillow
opencv-python-headless
numpy